requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
sqlalchemy
psycopg2-binary
python-jose[cryptography]
bcrypt
pulp
